+++ examples/scripts/battery_parameterisation/simple_ecm.py
@@ -69,12 +69,10 @@ true_values = [parameter_values[p] for p in ["R0 [Ohm]", "R1 [Ohm]"]]
 parameter_values.update(
     {
         "R0 [Ohm]": pybop.Parameter(
-            "R0 [Ohm]",
             prior=pybop.Gaussian(0.0002, 0.0001),
             bounds=[1e-4, 1e-2],
         ),
         "R1 [Ohm]": pybop.Parameter(
-            "R1 [Ohm]",
             prior=pybop.Gaussian(0.0001, 0.0001),
             bounds=[1e-5, 1e-3],
         ),