+++ examples/scripts/getting_started/maximum_likelihood.py
@@ -36,12 +36,10 @@ parameter_values.update(
 parameter_values.update(
     {
         "Negative electrode active material volume fraction": pybop.Parameter(
-            "Negative electrode active material volume fraction",
             prior=pybop.Gaussian(0.6, 0.05),
             bounds=[0.5, 0.8],
         ),
         "Positive electrode active material volume fraction": pybop.Parameter(
-            "Positive electrode active material volume fraction",
             prior=pybop.Gaussian(0.48, 0.05),
         ),
     }