+++ examples/scripts/getting_started/monte_carlo_sampling.py
@@ -44,12 +44,10 @@ dataset = pybop.Dataset(
 parameter_values.update(
     {
         "Negative electrode active material volume fraction": pybop.Parameter(
-            "Negative electrode active material volume fraction",
             prior=pybop.Gaussian(0.68, 0.02),
             transformation=pybop.LogTransformation(),
         ),
         "Positive electrode active material volume fraction": pybop.Parameter(
-            "Positive electrode active material volume fraction",
             prior=pybop.Gaussian(0.65, 0.02),
             transformation=pybop.LogTransformation(),
         ),