+++ tests/integration/test_transformation.py
@@ -65,13 +65,11 @@ class TestTransformation:
     def parameters(self, transformation_r0, transformation_r1):
         return {
             "R0 [Ohm]": pybop.Parameter(
-                "R0 [Ohm]",
                 prior=pybop.Gaussian(0.05, 0.02),
                 bounds=[1e-4, 0.1],
                 transformation=transformation_r0,
             ),
             "R1 [Ohm]": pybop.Parameter(
-                "R1 [Ohm]",
                 prior=pybop.Gaussian(0.05, 0.02),
                 bounds=[1e-4, 0.1],
                 transformation=transformation_r1,