+++ tests/unit/test_plots.py
@@ -33,7 +33,6 @@ class TestPlots:
     def parameters(self):
         return {
             "Negative electrode active material volume fraction": pybop.Parameter(
-                "Negative electrode active material volume fraction",
                 prior=pybop.Gaussian(0.68, 0.05),
                 bounds=[0.5, 0.8],
                 transformation=pybop.ScaledTransformation(
@@ -41,7 +40,6 @@ class TestPlots:
                 ),
             ),
             "Positive electrode active material volume fraction": pybop.Parameter(
-                "Positive electrode active material volume fraction",
                 prior=pybop.Gaussian(0.58, 0.05),
                 bounds=[0.4, 0.7],
                 transformation=pybop.ScaledTransformation(
@@ -199,7 +197,6 @@ class TestPlots:
         parameter_values.update(
             {
                 "Negative electrode active material volume fraction": pybop.Parameter(
-                    "Negative electrode active material volume fraction",
                     prior=pybop.Gaussian(0.68, 0.05),
                     bounds=[0.5, 0.8],
                 ),
@@ -219,12 +216,10 @@ class TestPlots:
         parameter_values.update(
             {
                 "Positive electrode active material volume fraction": pybop.Parameter(
-                    "Positive electrode active material volume fraction",
                     prior=pybop.Gaussian(0.58, 0.05),
                     bounds=[0.4, 0.7],
                 ),
                 "Positive particle radius [m]": pybop.Parameter(
-                    "Positive particle radius [m]",
                     prior=pybop.Gaussian(4.8e-06, 0.05e-06),
                     bounds=[4e-06, 6e-06],
                 ),
@@ -246,7 +241,6 @@ class TestPlots:
         parameter_values.update(
             {
                 "Positive electrode thickness [m]": pybop.Parameter(
-                    "Positive electrode thickness [m]",
                     prior=pybop.Gaussian(60e-6, 1e-6),
                     bounds=[10e-6, 80e-6],
                 )